requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python>=4.5.0
easyocr>=1.6.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
regex>=2023.0.0
nltk>=3.7.0
python-dateutil>=2.8.0
openpyxl>=3.0.0
gunicorn>=20.0.0
